+++ src/c_lib/map_gen/dragon.cpp
@@ -356,9 +356,9 @@ class L_System
         void set_tokens(char* tokens)
         {
             this->tokens = (char*)malloc((strlen(tokens)+1) * sizeof(char));
-            this->n_tokens = strlen(tokens);
+            this->n_tokens = (int)strlen(tokens);
             strcpy(this->tokens, tokens);
-            set_n_rules(strlen(this->tokens));
+            set_n_rules((int)strlen(this->tokens));
         }
 
         char* parse_rule(char* rule, int* token_id)
@@ -398,7 +398,7 @@ class L_System
             assert(this->rules[token_id] == NULL);
             assert(this->rule_lengths[token_id] == 0);
             this->rules[token_id] = new_rule;
-            this->rule_lengths[token_id] = strlen(new_rule);    // Invalid read of size 4
+            this->rule_lengths[token_id] = (int)strlen(new_rule);    // Invalid read of size 4
         }
 
         //returns length of next rule
